# apispec.yaml
openapi: 3.0.2
info:
  title: Todo API
  version: "1.0"
  contact:
    name: My name
    email: email@example.com
  description: This API allows user to keep track of their tasks
servers:
  - url: "http://localhost:5001/api"
    description: Development server
paths:
  /todos:
    get:
      description: Get a list of todos
      operationId: get-todos
      tags:
        - todos
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
        500:
          description: Internal Server Error
    post:
      description: Create a new todo
      operationId: post-todos
      tags:
        - todos
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        400:
          description: Bad Request
        500:
          description: Internal Server Error
  /todos/{todoId}:
    parameters:
      - schema:
          type: string
        name: todoId
        in: path
        required: true
    get:
      description: Retrieve the information of the todo with the matching todo ID.
      operationId: get-todos-todoId
      tags:
        - todos
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        404:
          description: Not Found
        500:
          description: Internal Server Error
    patch:
      description: Update an existing todo.
      operationId: patch-todos-todoId
      tags:
        - todos
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        404:
          description: Not Found
        500:
          description: Internal Server Error
    delete:
      description: Deletes a specific todo
      operationId: delete-todos-todoId
      tags:
        - todos
      responses:
        200:
          description: OK
        404:
          description: Not Found
        500:
          description: Internal Server Error
components:
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
      required:
        - id
        - description
tags:
  - name: todos